+++ allauth/account/tests.py
@@ -4,7 +4,6 @@ import json
 import uuid
 from datetime import timedelta
 
-import django
 from django import forms
 from django.conf import settings
 from django.contrib.auth.models import AbstractUser, AnonymousUser
@@ -14,6 +13,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.test.client import Client, RequestFactory
 from django.test.utils import override_settings
+from django.urls import reverse
 from django.utils.timezone import now
 
 from allauth.account.forms import BaseSignupForm, SignupForm
@@ -26,7 +26,6 @@ from allauth.tests import Mock, TestCase, patch
 from allauth.utils import get_user_model, get_username_max_length
 
 from . import app_settings
-from ..compat import is_authenticated, reverse
 from .adapter import get_adapter
 from .auth_backends import AuthenticationBackend
 from .signals import user_logged_out
@@ -185,7 +184,7 @@ class AccountTests(TestCase):
     def _create_user_and_login(self, usable_password=True):
         password = 'doe' if usable_password else False
         user = self._create_user(password=password)
-        self.client_force_login(user)
+        self.client.force_login(user)
         return user
 
     def test_redirect_when_authenticated(self):
@@ -208,6 +207,19 @@ class AccountTests(TestCase):
             resp,
             reverse('account_change_password'),
             fetch_redirect_response=False)
+
+    def test_set_password_not_allowed(self):
+        user = self._create_user_and_login(True)
+        pwd = '!*123i1uwn12W23'
+        self.assertFalse(user.check_password(pwd))
+        resp = self.client.post(
+            reverse('account_set_password'),
+            data={'password1': pwd,
+                  'password2': pwd})
+        user.refresh_from_db()
+        self.assertFalse(user.check_password(pwd))
+        self.assertTrue(user.has_usable_password())
+        self.assertEqual(resp.status_code, 302)
 
     def test_password_change_no_redirect(self):
         resp = self._password_set_or_change_redirect(
@@ -270,6 +282,36 @@ class AccountTests(TestCase):
         self.assertEqual(len(mail.outbox), 1)
         self.assertEqual(mail.outbox[0].to, ["john@example.org"])
         return user
+
+    def test_password_reset_flow_with_empty_session(self):
+        """
+        Test the password reset flow when the session is empty:
+        requesting a new password, receiving the reset link via email,
+        following the link, getting redirected to the
+        new link (without the token)
+        Copying the link and using it in a DIFFERENT client (Browser/Device).
+        """
+        # Request new password
+        self._request_new_password()
+        body = mail.outbox[0].body
+        self.assertGreater(body.find('https://'), 0)
+
+        # Extract URL for `password_reset_from_key` view
+        url = body[body.find('/password/reset/'):].split()[0]
+        resp = self.client.get(url)
+
+        reset_pass_url = resp.url
+
+        # Accesing the url via a different session
+        resp = self.client_class().get(reset_pass_url)
+
+        # We should receive the token_fail context_data
+        self.assertTemplateUsed(
+            resp,
+            'account/password_reset_from_key.%s' %
+            app_settings.TEMPLATE_EXTENSION)
+
+        self.assertTrue(resp.context_data['token_fail'])
 
     def test_password_reset_flow(self):
         """
@@ -336,6 +378,26 @@ class AccountTests(TestCase):
         data = json.loads(response.content.decode('utf8'))
         assert 'invalid' in data['form']['errors'][0]
 
+    def test_password_reset_flow_with_email_changed(self):
+        """
+        Test that the password reset token is invalidated if
+        the user email address was changed.
+        """
+        user = self._request_new_password()
+        body = mail.outbox[0].body
+        self.assertGreater(body.find('https://'), 0)
+        EmailAddress.objects.create(
+            user=user,
+            email='other@email.org')
+        # Extract URL for `password_reset_from_key` view
+        url = body[body.find('/password/reset/'):].split()[0]
+        resp = self.client.get(url)
+        self.assertTemplateUsed(
+            resp,
+            'account/password_reset_from_key.%s' %
+            app_settings.TEMPLATE_EXTENSION)
+        self.assertTrue('token_fail' in resp.context_data)
+
     @override_settings(ACCOUNT_LOGIN_ON_PASSWORD_RESET=True)
     def test_password_reset_ACCOUNT_LOGIN_ON_PASSWORD_RESET(self):
         user = self._request_new_password()
@@ -348,7 +410,7 @@ class AccountTests(TestCase):
             resp.url,
             {'password1': 'newpass123',
              'password2': 'newpass123'})
-        self.assertTrue(is_authenticated(user))
+        self.assertTrue(user.is_authenticated)
         # EmailVerificationMethod.MANDATORY sends us to the confirm-email page
         self.assertRedirects(resp, '/confirm-email/')
 
@@ -643,8 +705,6 @@ class AccountTests(TestCase):
         }
     }])
     def test_django_password_validation(self):
-        if django.VERSION < (1, 9, ):
-            return
         resp = self.client.post(
             reverse('account_signup'),
             {'username': 'johndoe',
@@ -973,11 +1033,13 @@ class AuthenticationBackendTests(TestCase):
         backend = AuthenticationBackend()
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.username,
                 password=user.username).pk,
             user.pk)
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.email,
                 password=user.username),
             None)
@@ -989,11 +1051,13 @@ class AuthenticationBackendTests(TestCase):
         backend = AuthenticationBackend()
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.email,
                 password=user.username).pk,
             user.pk)
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.username,
                 password=user.username),
             None)
@@ -1005,11 +1069,13 @@ class AuthenticationBackendTests(TestCase):
         backend = AuthenticationBackend()
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.email,
                 password=user.username).pk,
             user.pk)
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.username,
                 password=user.username).pk,
             user.pk)