+++ allauth/socialaccount/providers/draugiem/provider.py
@@ -1,6 +1,6 @@
+from django.urls import reverse
 from django.utils.http import urlencode
 
-from allauth.compat import reverse
 from allauth.socialaccount.providers.base import Provider, ProviderAccount