+++ allauth/socialaccount/providers/dwolla/urls.py
@@ -1,5 +1,3 @@
-"""URLs for Dwolla Provider"""
-
 from allauth.socialaccount.providers.oauth2.urls import default_urlpatterns
 
 from .provider import DwollaProvider