+++ allauth/socialaccount/providers/dwolla/views.py
@@ -1,5 +1,3 @@
-"""Views for DwollaProvider"""
-
 import requests
 
 from django.conf import settings