+++ allauth/socialaccount/providers/oauth2/provider.py
@@ -1,6 +1,7 @@
+from django.urls import reverse
 from django.utils.http import urlencode
 
-from allauth.compat import parse_qsl, reverse
+from allauth.compat import parse_qsl
 from allauth.socialaccount.providers.base import Provider