+++ allauth/socialaccount/providers/oauth2/views.py
@@ -5,9 +5,9 @@ from requests import RequestException
 
 from django.core.exceptions import PermissionDenied
 from django.http import HttpResponseRedirect
+from django.urls import reverse
 from django.utils import timezone
 
-from allauth.compat import reverse
 from allauth.exceptions import ImmediateHttpResponse
 from allauth.socialaccount import providers
 from allauth.socialaccount.helpers import (