+++ allauth/tests.py
@@ -5,13 +5,10 @@ import json
 import requests
 from datetime import date, datetime
 
-import django
 from django.core.files.base import ContentFile
 from django.db import models
-from django.test import TestCase as DjangoTestCase, override_settings, RequestFactory
-from django.http import HttpRequest
+from django.test import TestCase, override_settings, RequestFactory
 
-from allauth.account.utils import user_username
 from allauth.compat import base36_to_int, int_to_base36
 
 from . import utils
@@ -21,33 +18,6 @@ try:
     from unittest.mock import Mock, patch
 except ImportError:
     from mock import Mock, patch  # noqa
-
-
-class TestCase(DjangoTestCase):
-
-    def assertRedirects(self, response, expected_url,
-                        fetch_redirect_response=True,
-                        **kwargs):
-        super(TestCase, self).assertRedirects(
-            response,
-            expected_url,
-            fetch_redirect_response=fetch_redirect_response,
-            **kwargs)
-
-    def client_force_login(self, user):
-        if django.VERSION >= (1, 9):
-            self.client.force_login(
-                user,
-                'django.contrib.auth.backends.ModelBackend')
-        else:
-            old_password = user.password
-            user.set_password('doe')
-            user.save()
-            self.client.login(
-                username=user_username(user),
-                password='doe')
-            user.password = old_password
-            user.save()
 
 
 class MockedResponse(object):